requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
defusedxml==0.7.1
feedparser==6.0.12
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pytube==15.0.0
requests==2.32.5
sgmllib3k==1.0.0
urllib3==2.6.3
webvtt-py==0.5.1
Werkzeug==3.1.5
youtube-transcript-api==0.6.2
yt-dlp==2025.12.8
